Build Backend and Implement
Build a backend with typespecs, then implement the behaviour in a
module for tying into WolframAlpha as a backend. The Wolfram module
sends a query string to WolframAlpha, which responds with XML for
which SweetXml was added as a dependency for handling. A developer
ID was created with WolframAlpha and stored in an unversioned file
for dev, but would be setup as an ENV variable when deployed. The
results are built into a InfoSys.Result struct, which is defined in
the InfoSys module.

Backend calls were built to be accessible by any number of configured
backend services. They are supervised, incorporate timeouts enforced by
tasks, implement caching, sort results on relevance, and return up to
a specified number of results