Integrate Information System With Channels
Add the information system as a dependency of RumblWeb, then update
the VideoChannel to leverage InfoSys. When annotating a video the
annotation is immediately broadcast to all users in the topic, but
the information system has ten seconds to take the annotation as a
query and respond with relevant information. If there is a return, it
is posted as an annotation by the backend name. Also added a file for
seeding the User table with backend users