Test Channels
Build helpers for creating default users and videos, inserting users
and videos, and logging in users. Write tests for the UserSocket to
ensure valid authentication is required. Test the VideoChannel to
ensure logging in replies with annotations, creating annotations
causes them to broadcast, and new annotations trigger the
information system